+++ Tests/Test.Eyes.Sdk.Core.DotNet/TestAPI.cs
@@ -23,16 +23,19 @@ namespace Applitools.Tests
             PropertyInfo[] properties = configType.GetProperties(BindingFlags.Public | BindingFlags.Instance);
             foreach (PropertyInfo pi in properties)
             {
-                Assert.NotNull(pi.GetSetMethod());
-                Assert.NotNull(pi.GetGetMethod());
+                Assert.NotNull(pi.GetSetMethod(), $"Property {pi.Name} doesn't have SET method");
+                Assert.NotNull(pi.GetGetMethod(), $"Property {pi.Name} doesn't have GET method");
                 {
                     MethodInfo mi = configType.GetMethod("Set" + pi.Name, BindingFlags.Public | BindingFlags.Instance);
                     Assert.NotNull(mi, "property '{0}' doesn't have matching setter", pi.Name);
                     Assert.AreEqual(configType, mi.ReturnType);
                     ParameterInfo[] paramsInfo = mi.GetParameters();
                     Assert.AreEqual(1, paramsInfo.Length);
-                    Assert.AreEqual(pi.PropertyType, paramsInfo[0].ParameterType);
+                    Assert.IsTrue(pi.PropertyType.IsAssignableFrom(paramsInfo[0].ParameterType),
+                        "Setter method parameter type {0} is not assignable from {1}", paramsInfo[0].ParameterType, pi.PropertyType);
+                    //Assert.AreEqual(pi.PropertyType, paramsInfo[0].ParameterType);
                 }
+
                 {
                     MethodInfo mi = seleniumConfigType.GetMethod("Set" + pi.Name, BindingFlags.Public | BindingFlags.Instance);
                     Assert.NotNull(mi, "method 'Set{0}' isn't overriden in Selenium.Configuration", pi.Name);
@@ -44,7 +47,8 @@ namespace Applitools.Tests
                     Assert.AreEqual(seleniumConfigType, mi.ReturnType);
                     ParameterInfo[] paramsInfo = mi.GetParameters();
                     Assert.AreEqual(1, paramsInfo.Length);
-                    Assert.AreEqual(pi.PropertyType, paramsInfo[0].ParameterType);
+                    Assert.IsTrue(pi.PropertyType.IsAssignableFrom(paramsInfo[0].ParameterType),
+                        "Setter method parameter type {0} is not assignable from {1}", paramsInfo[0].ParameterType, pi.PropertyType);
                 }
             }
             TestContext.Progress.WriteLine("{0} properties has set methods.", properties.Length);
@@ -360,6 +364,15 @@ namespace Applitools.Tests
                     list.Add(new VisualGridOption("option2", true));
                 }
                 modifiedValue = list.ToArray();
+            }
+            else if (type == typeof(IList<int>))
+            {
+                var list = new int[] { 1, 2, 3 };
+                if (origValue != null && list.Length == ((IList<int>)origValue).Count)
+                {
+                    list = new int[] { 1, 2, 3, 4 };
+                }
+                modifiedValue = list;
             }
             return modifiedValue;
         }